Refactor IoStream into own module
