+++ src/forward_proxy.rs
@@ -1,66 +1,17 @@
 use crate::compression::Compressor;
 use crate::errors::*;
+use crate::iostream::IoStream;
+use crate::reverse_proxy;
 use nix::sys::socket;
 use std::io::Write;
 use std::net::SocketAddr;
 use std::os::unix::io::AsRawFd;
-use tokio::io::{
-    split, AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt, ReadBuf, ReadHalf, WriteHalf,
-};
-use tokio::net::{TcpListener, TcpStream};
-use tokio_rustls::{client::TlsStream, rustls::ClientConfig, webpki::DNSNameRef, TlsConnector};
-
-use crate::reverse_proxy;
-use std::pin::Pin;
 use std::sync::Arc;
-use std::task::{Context, Poll};
+use tokio::io::{split, AsyncReadExt, AsyncWriteExt, ReadHalf, WriteHalf};
+use tokio::net::{TcpListener, TcpStream};
+use tokio_rustls::{rustls::ClientConfig, webpki::DNSNameRef, TlsConnector};
 
 pub const PROXY_REDIR_PORT: u16 = 8080;
-
-enum IoStream {
-    TcpStream(TcpStream),
-    TlsStream(TlsStream<TcpStream>),
-}
-
-impl AsyncRead for IoStream {
-    fn poll_read(
-        self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &mut ReadBuf<'_>,
-    ) -> Poll<std::io::Result<()>> {
-        match self.get_mut() {
-            IoStream::TcpStream(stream) => Pin::new(stream).poll_read(cx, buf),
-            IoStream::TlsStream(stream) => Pin::new(stream).poll_read(cx, buf),
-        }
-    }
-}
-
-impl AsyncWrite for IoStream {
-    fn poll_write(
-        self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &[u8],
-    ) -> Poll<std::io::Result<usize>> {
-        match self.get_mut() {
-            IoStream::TcpStream(stream) => Pin::new(stream).poll_write(cx, buf),
-            IoStream::TlsStream(stream) => Pin::new(stream).poll_write(cx, buf),
-        }
-    }
-
-    fn poll_flush(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<std::io::Result<()>> {
-        match self.get_mut() {
-            IoStream::TcpStream(stream) => Pin::new(stream).poll_flush(cx),
-            IoStream::TlsStream(stream) => Pin::new(stream).poll_flush(cx),
-        }
-    }
-
-    fn poll_shutdown(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<std::io::Result<()>> {
-        match self.get_mut() {
-            IoStream::TcpStream(stream) => Pin::new(stream).poll_shutdown(cx),
-            IoStream::TlsStream(stream) => Pin::new(stream).poll_shutdown(cx),
-        }
-    }
-}
 
 pub fn run(local_addr: SocketAddr, compress: bool, encrypt: bool) -> Result<()> {
     let rt = tokio::runtime::Runtime::new().chain_err(|| "failed to create tokio runtime")?;
